wip: 🚧 project detail
