+++ src/components/Sider/index.tsx
@@ -9,7 +9,9 @@ import s from './index.module.less'
 import { DEFAULT_SETTINGS } from '@/util/constants'
 import { getInitalSettings } from '@/util/baseInfo'
 
-const index: React.FC<{}> = () => {
+const index: React.FC<{
+  defaultRoute?: string
+}> = ({ defaultRoute }) => {
   const navigate = useNavigate()
   const { homePage = DEFAULT_SETTINGS.homePage } = getInitalSettings()
 
@@ -17,6 +19,10 @@ const index: React.FC<{}> = () => {
     const { path } = MENUS.find(item => item.value === value) || { path: '/' }
     navigate(path)
   }
+
+  useEffect(() => {
+    if (defaultRoute) navigate(defaultRoute)
+  }, [defaultRoute])
 
   return (
     <div className={s.container}>
