+++ src/pages/Calendar/index.module.less
@@ -3,14 +3,6 @@
   @apply h-screen;
 }
 
-.content {
-
-  .calendarWrapper {
-    background-color: var(--ls-quaternary-background-color);
-  }
-
-}
-
 .sideBar {
   width: 200px;
   @apply h-screen border-0 border-l border-l-gray-200 border-solid;
