+++ src/util/util.ts
@@ -125,3 +125,12 @@ export const extractDays = (startDate: Dayjs, endDate: Dayjs): Dayjs[] => {
   }
   return days
 }
+
+export const genRandomString = (length: number = 6) => {
+  const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+  let result = ''
+  for (let i = 0; i < length; i++) {
+    result += chars.charAt(Math.floor(Math.random() * chars.length))
+  }
+  return result
+}
